# frontend/contracts/openapi/ops-api.yaml
...
openapi: 3.1.1
info:
  title: AI Sales Operating System Backend API (Operations)
  version: 2.0.0
  description: Contract starter generated from Backend Enterprise Implementation
    Blueprint v2.0. Each sprint MUST refine operation-specific schemas/examples
    before freeze.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: https://api.staging.example.com/api/v1
    description: Staging
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for browser requests authenticated by the BFF session
        cookie. The value must match the non-HttpOnly csrf_token cookie
        (double-submit). Not required for OAuth/PKCE bearer clients. Missing or
        mismatched values return 403 CSRF_TOKEN_INVALID.
      schema:
        type: string
        minLength: 32
        maxLength: 256
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  headers:
    XRequestId:
      schema:
        type: string
      description: Server request ID
    XCorrelationId:
      schema:
        type: string
      description: Business correlation ID
...
  schemas:
    Meta:
      type: object
      additionalProperties: false
      required:
        - request_id
      properties:
        request_id:
          type: string
        correlation_id:
          type:
            - string
            - "null"
...
    FieldError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        field:
          type:
            - string
            - "null"
        code:
          type: string
        message:
          type: string
        meta:
          type: object
          additionalProperties: true
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - code
        - request_id
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type:
            - string
            - "null"
        instance:
          type:
            - string
            - "null"
        request_id:
          type: string
        correlation_id:
          type:
            - string
            - "null"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
...
    OperationsResource:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        tenant_id:
          type:
            - string
            - "null"
          format: uuid
        created_at:
          type: string
          format: date-time
        detail:
          type:
            - object
            - "null"
          additionalProperties: true
      description: Frozen resource schema for OpenAPI tag Operations (enterprise
        doc-freeze W1).
...
    CreateSupportAccessRequest:
      type: object
      additionalProperties: false
      required:
        - tenant_id
        - expires_at
      properties:
        tenant_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
        reason:
          type: string
          minLength: 1
          maxLength: 1000
      description: Request body for createSupportAccess (W1 freeze).
    CreateSupportAccessResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: "#/components/schemas/OperationsResource"
        meta:
          $ref: "#/components/schemas/Meta"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/validation-failed
            title: Bad request
            status: 400
            code: VALIDATION_FAILED
            request_id: req_example
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/auth-invalid-credentials
            title: Authentication required
            status: 401
            code: AUTH_INVALID_CREDENTIALS
            request_id: req_example
    Forbidden:
      description: Permission denied
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/insufficient-permission
            title: Permission denied
            status: 403
            code: INSUFFICIENT_PERMISSION
            request_id: req_example
    NotFound:
      description: Resource not found or not visible
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/resource-not-found
            title: Resource not found or not visible
            status: 404
            code: RESOURCE_NOT_FOUND
            request_id: req_example
    Conflict:
      description: State, duplicate, or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/resource-state-conflict
            title: State, duplicate, or idempotency conflict
            status: 409
            code: RESOURCE_STATE_CONFLICT
            request_id: req_example
    PreconditionFailed:
      description: ETag/precondition mismatch
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/resource-version-mismatch
            title: ETag/precondition mismatch
            status: 412
            code: RESOURCE_VERSION_MISMATCH
            request_id: req_example
    Unprocessable:
      description: Business validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/validation-failed
            title: Business validation failed
            status: 422
            code: VALIDATION_FAILED
            request_id: req_example
    TooManyRequests:
      description: Rate or quota exceeded
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/rate-limited
            title: Rate or quota exceeded
            status: 429
            code: RATE_LIMITED
            request_id: req_example
    ServerError:
      description: Unexpected server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/internal-error
            title: Unexpected server error
            status: 500
            code: INTERNAL_ERROR
            request_id: req_example
    ServiceUnavailable:
      description: Dependency unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.example.com/service-unavailable
            title: Dependency unavailable
            status: 503
            code: SERVICE_UNAVAILABLE
            request_id: req_example
paths:
...
  /super-admin/tenants/{tenant_id}/support-access:
    post:
      tags:
        - Operations
      summary: Create support access
      operationId: createSupportAccess
      x-permission: support.access
      x-idempotency: required
      x-csrf-protection: cookie-session-required
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "201":
          description: Success
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-Correlation-Id:
              $ref: "#/components/headers/XCorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSupportAccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportAccessRequest"
...